+++ SentryTestUtils/Sources/ClearTestState.swift
@@ -31,7 +31,6 @@ class TestCleanup: NSObject {
         SentrySDK.close()
         SentrySDKInternal.setCurrentHub(nil)
         SentrySDKInternal.lastRunStatusCalled = false
-        SentrySDKInternal.crashReporterInstalled = false
         SentrySDKInternal.fatalDetected = false
         SentrySDKInternal.startInvocations = 0
         SentrySDKInternal.setDetectedStartUpCrash(false)